make program run
